source/requirements.txt
Django
geopandas
GDAL
psycopg2-binary
python-decouple